1st Working Model
This model uses a Backwards-Euler time integration method and solves the non-linear system of equations using a modified Newton-Raphson scheme. This code will be the base to build up from and will be used to develop further in terms of adding adaptive step sizes and initialzing the model within a class variable.